// packages/shared/src/profile/availability.enums.ts
export enum AvailabilityFrequency {
  HOURS_WEEK = "HOURS_WEEK",
  HOURS_MONTH = "HOURS_MONTH",
  DAYS_WEEK = "DAYS_WEEK",
  DAYS_MONTH = "DAYS_MONTH",
  ONE_DAY = "ONE_DAY",
  PUNCTUAL = "PUNCTUAL",
}

export enum AvailabilityTime {
  WEEKDAY = "WEEKDAY",
  WEEKEND = "WEEKEND",
  EVENING = "EVENING",
  ALL_TIME = "ALL_TIME",
}

export enum AvailabilityType {
  REMOTE = "REMOTE",
  ON_SITE = "ON_SITE",
  HYBRID = "HYBRID",
}